+++ src/index.js
@@ -3,5 +3,5 @@
 import update from './update';
 import wrapCursor from './wrapCursor';
 
-module.exports = update;
-(module.exports: any).wrapCursor = wrapCursor;
+module.exports.update = update;
+module.exports.wrapCursor = wrapCursor;