+++ CHANGELOG.md
@@ -7,11 +7,16 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This change log adheres to [keepachangelog.com](http://keepachangelog.com).
 
 ## [Unreleased]
+### Fixed
+- Handle inactive textarea element
+- Enable to insert a string similar to texts around the cursor
 
 ## [0.1.1] - 2017-06-28
+### Fixed
 - ES2015 compatible
 
 ## 0.1.0 - 2017-06-28
+### Added
 - Initial release
 
 [Unreleased]: https://github.com/yuku-t/undate/compare/v0.1.1...HEAD